Move AppState to separate file
