+++ src/interactive/app/mod.rs
@@ -6,6 +6,7 @@ pub mod input;
 mod navigation;
 pub mod tree_view;
 pub mod terminal_app;
+pub mod app_state;
 
 pub use bytevis::*;
 pub use common::*;
